depository: split into smaller files
