+++ internal/depository/micro_deposits_test.go
@@ -629,61 +629,6 @@ func TestMicroDeposits__MarkMicroDepositAsMerged(t *testing.T) {
 	check(t, NewDepositoryRepo(log.NewNopLogger(), mysqlDB.DB, keeper))
 }
 
-func TestMicroDepositCursor__next(t *testing.T) {
-	sqliteDB := database.CreateTestSqliteDB(t)
-	defer sqliteDB.Close()
-
-	keeper := secrets.TestStringKeeper(t)
-
-	depRepo := NewDepositoryRepo(log.NewNopLogger(), sqliteDB.DB, keeper)
-	cur := depRepo.GetMicroDepositCursor(2)
-
-	microDeposits, err := cur.Next()
-	if len(microDeposits) != 0 || err != nil {
-		t.Fatalf("microDeposits=%#v error=%v", microDeposits, err)
-	}
-
-	// Write a micro-deposit
-	amt, _ := model.NewAmount("USD", "0.11")
-	if err := depRepo.InitiateMicroDeposits(id.Depository("id"), "userID", []*MicroDeposit{{Amount: *amt, FileID: "fileID"}}); err != nil {
-		t.Fatal(err)
-	}
-	// our cursor should return this micro-deposit now since there's no mergedFilename
-	microDeposits, err = cur.Next()
-	if len(microDeposits) != 1 || err != nil {
-		t.Fatalf("microDeposits=%#v error=%v", microDeposits, err)
-	}
-	if microDeposits[0].DepositoryID != "id" || microDeposits[0].Amount.String() != "USD 0.11" {
-		t.Errorf("microDeposits[0]=%#v", microDeposits[0])
-	}
-	mc := microDeposits[0] // save for later
-
-	// verify calling our cursor again returns nothing
-	microDeposits, err = cur.Next()
-	if len(microDeposits) != 0 || err != nil {
-		t.Fatalf("microDeposits=%#v error=%v", microDeposits, err)
-	}
-
-	// mark the micro-deposit as merged (via merged_filename) and re-create the cursor to expect nothing returned in Next()
-	cur = depRepo.GetMicroDepositCursor(2)
-	if err := depRepo.MarkMicroDepositAsMerged("filename", mc); err != nil {
-		t.Fatal(err)
-	}
-	microDeposits, err = cur.Next()
-	if len(microDeposits) != 0 || err != nil {
-		t.Fatalf("microDeposits=%#v error=%v", microDeposits, err)
-	}
-
-	// verify merged_filename
-	filename, err := ReadMergedFilename(depRepo, mc.Amount, id.Depository(mc.DepositoryID))
-	if err != nil {
-		t.Fatal(err)
-	}
-	if filename != "filename" {
-		t.Errorf("mc=%#v", mc)
-	}
-}
-
 func TestMicroDeposits__addMicroDeposit(t *testing.T) {
 	amt, _ := model.NewAmount("USD", "0.28")
