+++ internal/depository/repository.go
@@ -1,0 +1,250 @@
+// Copyright 2020 The Moov Authors
+// Use of this source code is governed by an Apache License
+// license that can be found in the LICENSE file.
+
+package depository
+
+import (
+	"database/sql"
+	"fmt"
+	"time"
+
+	"github.com/go-kit/kit/log"
+	"github.com/moov-io/base"
+	"github.com/moov-io/paygate/internal/database"
+	"github.com/moov-io/paygate/internal/hash"
+	"github.com/moov-io/paygate/internal/model"
+	"github.com/moov-io/paygate/internal/secrets"
+	"github.com/moov-io/paygate/pkg/id"
+)
+
+type Repository interface {
+	GetDepository(id id.Depository) (*model.Depository, error) // admin endpoint
+	GetUserDepositories(userID id.User) ([]*model.Depository, error)
+	GetUserDepository(id id.Depository, userID id.User) (*model.Depository, error)
+
+	UpsertUserDepository(userID id.User, dep *model.Depository) error
+	UpdateDepositoryStatus(id id.Depository, status model.DepositoryStatus) error
+	deleteUserDepository(id id.Depository, userID id.User) error
+
+	GetMicroDeposits(id id.Depository) ([]*MicroDeposit, error) // admin endpoint
+	getMicroDepositsForUser(id id.Depository, userID id.User) ([]*MicroDeposit, error)
+
+	LookupDepositoryFromReturn(routingNumber string, accountNumber string) (*model.Depository, error)
+	LookupMicroDepositFromReturn(id id.Depository, amount *model.Amount) (*MicroDeposit, error)
+	SetReturnCode(id id.Depository, amount model.Amount, returnCode string) error
+
+	InitiateMicroDeposits(id id.Depository, userID id.User, microDeposit []*MicroDeposit) error
+	confirmMicroDeposits(id id.Depository, userID id.User, amounts []model.Amount) error
+	GetMicroDepositCursor(batchSize int) *MicroDepositCursor
+}
+
+func NewDepositoryRepo(logger log.Logger, db *sql.DB, keeper *secrets.StringKeeper) *SQLRepo {
+	return &SQLRepo{logger: logger, db: db, keeper: keeper}
+}
+
+type SQLRepo struct {
+	db     *sql.DB
+	logger log.Logger
+	keeper *secrets.StringKeeper
+}
+
+func (r *SQLRepo) Close() error {
+	return r.db.Close()
+}
+
+func (r *SQLRepo) GetDepository(depID id.Depository) (*model.Depository, error) {
+	query := `select user_id from depositories where depository_id = ? and deleted_at is null limit 1;`
+	stmt, err := r.db.Prepare(query)
+	if err != nil {
+		return nil, err
+	}
+	defer stmt.Close()
+
+	var userID string
+	if err := stmt.QueryRow(depID).Scan(&userID); err != nil {
+		if err == sql.ErrNoRows {
+			return nil, nil
+		}
+		return nil, err
+	}
+	if userID == "" {
+		return nil, nil // not found
+	}
+
+	dep, err := r.GetUserDepository(depID, id.User(userID))
+	if err == sql.ErrNoRows {
+		return nil, nil
+	}
+	return dep, err
+}
+
+func (r *SQLRepo) GetUserDepositories(userID id.User) ([]*model.Depository, error) {
+	query := `select depository_id from depositories where user_id = ? and deleted_at is null`
+	stmt, err := r.db.Prepare(query)
+	if err != nil {
+		return nil, err
+	}
+	defer stmt.Close()
+
+	rows, err := stmt.Query(userID)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+
+	var depositoryIds []string
+	for rows.Next() {
+		var row string
+		if err := rows.Scan(&row); err != nil {
+			return nil, fmt.Errorf("GetUserDepositories scan: %v", err)
+		}
+		if row != "" {
+			depositoryIds = append(depositoryIds, row)
+		}
+	}
+
+	var depositories []*model.Depository
+	for i := range depositoryIds {
+		dep, err := r.GetUserDepository(id.Depository(depositoryIds[i]), userID)
+		if err == nil && dep != nil && dep.BankName != "" {
+			depositories = append(depositories, dep)
+		}
+	}
+	return depositories, rows.Err()
+}
+
+func (r *SQLRepo) GetUserDepository(id id.Depository, userID id.User) (*model.Depository, error) {
+	query := `select depository_id, bank_name, holder, holder_type, type, routing_number, account_number_encrypted, account_number_hashed, status, metadata, created_at, last_updated_at
+from depositories
+where depository_id = ? and user_id = ? and deleted_at is null
+limit 1`
+	stmt, err := r.db.Prepare(query)
+	if err != nil {
+		return nil, fmt.Errorf("GetUserDepository: prepare: %v", err)
+	}
+	defer stmt.Close()
+
+	row := stmt.QueryRow(id, userID)
+
+	dep := &model.Depository{UserID: userID}
+	var (
+		created time.Time
+		updated time.Time
+	)
+	err = row.Scan(&dep.ID, &dep.BankName, &dep.Holder, &dep.HolderType, &dep.Type, &dep.RoutingNumber, &dep.EncryptedAccountNumber, &dep.HashedAccountNumber, &dep.Status, &dep.Metadata, &created, &updated)
+	if err != nil {
+		if err == sql.ErrNoRows {
+			return nil, nil
+		}
+		return nil, fmt.Errorf("GetUserDepository: scan: %v", err)
+	}
+	dep.ReturnCodes = r.getMicroDepositReturnCodes(dep.ID)
+	dep.Created = base.NewTime(created)
+	dep.Updated = base.NewTime(updated)
+	if dep.ID == "" || dep.BankName == "" {
+		return nil, nil // no records found
+	}
+	dep.Keeper = r.keeper
+	return dep, nil
+}
+
+func (r *SQLRepo) UpsertUserDepository(userID id.User, dep *model.Depository) error {
+	tx, err := r.db.Begin()
+	if err != nil {
+		return err
+	}
+
+	now := base.NewTime(time.Now())
+	if dep.Created.IsZero() {
+		dep.Created = now
+		dep.Updated = now
+	}
+
+	query := `insert into depositories (depository_id, user_id, bank_name, holder, holder_type, type, routing_number, account_number_encrypted, account_number_hashed, status, metadata, created_at, last_updated_at)
+values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?);`
+	stmt, err := tx.Prepare(query)
+	if err != nil {
+		return err
+	}
+	defer stmt.Close()
+
+	res, err := stmt.Exec(dep.ID, userID, dep.BankName, dep.Holder, dep.HolderType, dep.Type, dep.RoutingNumber, dep.EncryptedAccountNumber, dep.HashedAccountNumber, dep.Status, dep.Metadata, dep.Created.Time, dep.Updated.Time)
+	stmt.Close()
+	if err != nil && !database.UniqueViolation(err) {
+		return fmt.Errorf("problem upserting depository=%q, userID=%q: %v", dep.ID, userID, err)
+	}
+	if res != nil {
+		if n, _ := res.RowsAffected(); n != 0 {
+			return tx.Commit() // Depository was inserted, so cleanup and exit
+		}
+	}
+	query = `update depositories
+set bank_name = ?, holder = ?, holder_type = ?, type = ?, routing_number = ?,
+account_number_encrypted = ?, account_number_hashed = ?, status = ?, metadata = ?, last_updated_at = ?
+where depository_id = ? and user_id = ? and deleted_at is null`
+	stmt, err = tx.Prepare(query)
+	if err != nil {
+		return err
+	}
+	defer stmt.Close()
+	_, err = stmt.Exec(
+		dep.BankName, dep.Holder, dep.HolderType, dep.Type, dep.RoutingNumber,
+		dep.EncryptedAccountNumber, dep.HashedAccountNumber, dep.Status, dep.Metadata, time.Now(), dep.ID, userID)
+	stmt.Close()
+	if err != nil {
+		return fmt.Errorf("UpsertUserDepository: exec error=%v rollback=%v", err, tx.Rollback())
+	}
+	return tx.Commit()
+}
+
+func (r *SQLRepo) UpdateDepositoryStatus(id id.Depository, status model.DepositoryStatus) error {
+	query := `update depositories set status = ?, last_updated_at = ? where depository_id = ? and deleted_at is null`
+	stmt, err := r.db.Prepare(query)
+	if err != nil {
+		return err
+	}
+	defer stmt.Close()
+
+	if _, err := stmt.Exec(status, time.Now(), id); err != nil {
+		return fmt.Errorf("error updating status depository_id=%q: %v", id, err)
+	}
+	return nil
+}
+
+func (r *SQLRepo) deleteUserDepository(id id.Depository, userID id.User) error {
+	query := `update depositories set deleted_at = ? where depository_id = ? and user_id = ? and deleted_at is null`
+	stmt, err := r.db.Prepare(query)
+	if err != nil {
+		return err
+	}
+	defer stmt.Close()
+
+	if _, err := stmt.Exec(time.Now(), id, userID); err != nil && err != sql.ErrNoRows {
+		return fmt.Errorf("error deleting depository_id=%q, user_id=%q: %v", id, userID, err)
+	}
+	return nil
+}
+
+func (r *SQLRepo) LookupDepositoryFromReturn(routingNumber string, accountNumber string) (*model.Depository, error) {
+	hash, err := hash.AccountNumber(accountNumber)
+	if err != nil {
+		return nil, err
+	}
+	// order by created_at to ignore older rows with non-null deleted_at's
+	query := `select depository_id, user_id from depositories where routing_number = ? and account_number_hashed = ? and deleted_at is null order by created_at desc limit 1;`
+	stmt, err := r.db.Prepare(query)
+	if err != nil {
+		return nil, err
+	}
+	defer stmt.Close()
+
+	depID, userID := "", ""
+	if err := stmt.QueryRow(routingNumber, hash).Scan(&depID, &userID); err != nil {
+		if err == sql.ErrNoRows {
+			return nil, nil
+		}
+		return nil, fmt.Errorf("LookupDepositoryFromReturn: %v", err)
+	}
+	return r.GetUserDepository(id.Depository(depID), id.User(userID))
+}
